macOS: move menu item validation into SurfaceView+MenuValidation.swift
Pure code motion out of Ghostty.SurfaceView.swift; the only change is
dropping the MARK header the file name now carries.
